refactor(integration): move GitHub-URL flow out of service.go into github_url.go
Extract AnalyzeFromGitHubURL, fetchAndValidateGitHubAnalysis,
validateRepoURLMatch, buildGitHubOnlyAnalysis, githubURLToPURL,
inferPURLFromLanguages, parseGitHubURL, generateVersionedPURL,
extractPackageManagersFromManifests, mapPackageManagerToEcosystem,
generatePURLsFromPackageManagers, and generatePURLForEcosystem (~430 lines)
into a new github_url.go sibling, matching the per-flow file convention
already established by github_url_batch.go, purl_batch.go, and the
populate_*/enrich_*/resolve_vanity.go files.

service.go now retains only the struct definition, options, constructor,
FetchAnalysis/FetchAnalysisWithGitHub, and shared converters
(createPackageFromPURL, buildVersionDetail, classifyAdvisory,
extractScorecardChecks). Zero signature changes; all methods remain on
*IntegrationService; all tests pass unchanged.
